chore(execution): split execute.c into three files
